shs/ui

+++ public/administrator/management/subject.js
@@ -38,12 +38,13 @@ const subjectTable = (level) => {
                             </td>
                             <td>
                                 <div class="btn-group" role="group" aria-label="Basic example">
-                                    <button type="button" style="font-size:9px" class="btn btn-sm btn-info pl-3 pr-3 editSubject editSub_${
-                                        val.id
-                                    }" id="${val.id}">Edit</button>
-                                    <button type="button" style="font-size:9px" class="btn btn-sm btn-danger deleteSubject deleteSub_${
+                                </button>
+                                    <button type="button" style="font-size:9px" class="btn btn-sm btn-danger text-white pb-2 pt-2 pl-3 pr-3 deleteSubject deleteSub_${
                                         val.id
                                     }" id="${val.id}">Delete</button>
+                                    <button type="button" style="font-size:9px" class="btn btn-sm btn-info text-white pb-2 pt-2 pl-3 pr-3 editSubject editSub_${
+                                        val.id
+                                    }" id="${val.id}">Edit</button>
                                 </div>
                             </td>
                         </tr>
@@ -61,6 +62,7 @@ const subjectTable = (level) => {
             getToast("error", "Eror", errorThrown);
         });
 };
+
 subjectTable(7);
 $("#subjectForm").submit(function (e) {
     e.preventDefault();
@@ -88,6 +90,7 @@ $("#subjectForm").submit(function (e) {
             document.getElementById("subjectForm").reset();
             $("input[name='id']").val("");
             $(".btnSaveSubject").html("Submit").attr("disabled", false);
+            getToast("success", "Successfully", "added new subject");
             cancelSubject.hide();
         })
         .fail(function (jqxHR, textStatus, errorThrown) {
@@ -118,11 +121,7 @@ $('input[name="subject_code"]').on("blur", function () {
         })
             .done(function (data) {
                 if (data) {
-                    getToast(
-                        "warning",
-                        "Warning",
-                        "This subject is already added"
-                    );
+                    getToast("warning", "Warning", "This subject is already added");
                     $('input[name="subject_code"]').addClass("is-invalid");
                 } else {
                     $('input[name="subject_code"]').removeClass("is-invalid");
@@ -151,13 +150,14 @@ $(document).on("click", ".editSubject", function () {
         .done(function (data) {
             cancelSubject.show();
             // console.log(data.id);
-            $(".editSub_" + id).html("Edit");
+            $(".editSub_" + id).html(`<i class="fas fa-edit"></i>`);
             $(".btnSaveSubject").html("Update");
             $("input[name='id']").val(data.id);
             $("select[name='grade_level']").val(data.grade_level);
             $("input[name='subject_code']").val(data.subject_code);
             $("input[name='descriptive_title']").val(data.descriptive_title);
             $("select[name='subject_for']").val(data.subject_for);
+            // getToast("success", "Success", "updated one subject");
         })
         .fail(function (jqxHR, textStatus, errorThrown) {
             console.log(jqxHR, textStatus, errorThrown);
@@ -166,25 +166,29 @@ $(document).on("click", ".editSubject", function () {
 });
 
 $(document).on("click", ".deleteSubject", function () {
-    let id = $(this).attr("id");
-    $.ajax({
-        url: "subject/delete/" + id,
-        type: "DELETE",
-        data: { _token: $('input[name="_token"]').val() },
-        beforeSend: function () {
-            $(".deleteSub_" + id).html(`
+    if (confirm("Are you sure,Do you want delete this?")) {
+        let id = $(this).attr("id");
+        $.ajax({
+            url: "subject/delete/" + id,
+            type: "DELETE",
+            data: { _token: $('input[name="_token"]').val() },
+            beforeSend: function () {
+                $(".deleteSub_" + id).html(`
             <div class="spinner-border spinner-border-sm" role="status">
                 <span class="sr-only">Loading...</span>
             </div>`);
-        },
-    })
-        .done(function (response) {
-            $(".deleteSub" + id).html("Delete");
-            getToast("success", "Success", "deleted one record");
-            subjectTable($("#selectedGL").val());
+            },
         })
-        .fail(function (jqxHR, textStatus, errorThrown) {
-            console.log(jqxHR, textStatus, errorThrown);
-            getToast("error", "Eror", errorThrown);
-        });
+            .done(function (response) {
+                $(".deleteSub" + id).html(`<i class="fas fa-times"></i>`);
+                getToast("warning", "Successfully", "deleted one subject");
+                subjectTable($("#selectedGL").val());
+            })
+            .fail(function (jqxHR, textStatus, errorThrown) {
+                console.log(jqxHR, textStatus, errorThrown);
+                getToast("error", "Eror", errorThrown);
+            });
+    } else {
+        return false;
+    }
 });
